pyproject.toml
[project]
name = "company-task-work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "ollama>=0.6.1",
    "openai>=2.30.0",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]
